backend/main-backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic==2.11.3
pydantic-settings==2.8.1

# LangGraph and LangChain - Compatible versions
langgraph==0.0.26
langchain==0.1.12
langchain-core==0.1.31
langchain-openai==0.0.8

# HTTP client for microservice communication
httpx==0.25.2
aiohttp==3.11.18

# Utilities
python-dotenv==1.1.0
typing-extensions==4.13.2
anyio==4.9.0
orjson==3.9.15

# Logging and monitoring
loguru==0.7.2
prometheus-fastapi-instrumentator==7.1.0

# Database and ORM
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Security and Authentication
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
bcrypt==4.1.2

# Timezone handling
pytz==2024.1

# Rate limiting
slowapi==0.1.9 